+++ unknown-note-server/src/main/java/unknownnote/unknownnoteserver/controller/DiaryController.java
@@ -5,13 +5,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import unknownnote.unknownnoteserver.dto.DiaryDTO;
 import unknownnote.unknownnoteserver.entity.Diary;
 import unknownnote.unknownnoteserver.service.DiaryService;
-import unknownnote.unknownnoteserver.service.JwtService;
+import unknownnote.unknownnoteserver.jwt.JwtService;
 import unknownnote.unknownnoteserver.service.ErrorService;
 
 import java.util.Arrays;
