+++ unknown-note-server/src/main/java/unknownnote/unknownnoteserver/jwt/JwtService.java
@@ -1,0 +1,30 @@
+package unknownnote.unknownnoteserver.jwt;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jws;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.security.Keys;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
+import java.nio.charset.StandardCharsets;
+
+import javax.crypto.SecretKey;
+import java.util.Date;
+
+@Service
+public class JwtService {
+
+    @Value("${spring.jwt.secret}")
+    private String SECRET_KEY; //비밀 키
+
+    public int getUserIdFromJwt(String jwtToken){
+        SecretKey key = Keys.hmacShaKeyFor(SECRET_KEY.getBytes(StandardCharsets.UTF_8));
+
+        return Jwts.parser()
+                .verifyWith(key)
+                .build().parseSignedClaims(jwtToken)
+                .getPayload().get("user_id", Integer.class);
+    }
+}
